8.4 video edit only by creator

+++ controllers/videoController.js
@@ -37,8 +37,11 @@ export const postVideoUpload = async (req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
       // console.log(newVideo);
     res.redirect(routes.videoDetail(newVideo.id));
     };
@@ -48,9 +51,9 @@ export const videoDetail = async (req, res) => {
         params: {id}
     } = req;
     try {
-        const clickedVideo = await Video.findById(id);
-    // console.log(clickedVideo);
-    res.render("videoDetail", {pageTitle: clickedVideo.title, clickedVideo });
+        const clickedVideo = await Video.findById(id).populate("creator");
+        // console.log(clickedVideo);
+        res.render("videoDetail", {pageTitle: clickedVideo.title, clickedVideo });
     } catch (error) {
         // console.log(error);
         res.redirect(routes.home);
@@ -63,7 +66,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const videoEdit = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${videoEdit.title}`, videoEdit });
+        if (videoEdit.creator != req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${videoEdit.title}`, videoEdit });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -87,7 +94,12 @@ export const deleteVideo = async (req, res) => {
         params: {id}
     } = req;
     try {
-        await Video.findOneAndRemove({ _id: id});
+        const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+        throw Error();
+    } else {
+        await Video.findOneAndRemove({ _id: id });
+    }
     } catch (error) {
         console.log(error);
     }
